Fix downloading crypto data, better unit tests

+++ src/main/java/com/newsscraper/data/WebDataFrame.java
@@ -2,14 +2,14 @@ package com.newsscraper.data;
 
 public class WebDataFrame {
     public String symbol;
-    public long timestamp;
+    public int timestamp;
     public float open;
     public float close;
     public float high;
     public float low;
     public float volume;
 
-    public WebDataFrame(String symbol, long timestamp, float open, float close, float high, float low, float volume) {
+    public WebDataFrame(String symbol, int timestamp, float open, float close, float high, float low, float volume) {
         this.symbol = symbol;
         this.timestamp = timestamp;
         this.open = open;
